Split out manager into its own code
